\name{datStand}
\alias{datStand}
\title{Data fo the Section 'Spectral standardization'}
\usage{
data('datStand')
}
\description{
A list with four objects:
\itemize{
   \item{CSIRO.s: benchmark LB spectra, data frame with two rows and 2151 columns.}
   \item{ISS_slaves: spectra collected from two instruments, data frame with two rows and 2151 columns.}
   \item{asd.soil: two soil samples scans by an ASD instrument, data frame with two rows and 2151 columns.}
   \item{se.soil: two soil samples scans by an SE instrument, data frame with two rows and 2151 columns.}
}}
\examples{
data('datStand')
}
