\name{mineralRef}
\alias{mineralRef}
\title{Reference spectra for some common clay minerals and iron oxides.}
\usage{
data('mineralRef')
}
\description{
This spectral library was compiled by the USGS, including not only soils but minerals, vegetables and others. In this package we used only a fraction of this dataset including the following clay minerals: kaolinite (KGa-2), illite (GDS4),smectite (SWy-1), kaolinite–smectite 50/50 mixture (H89-FR-2), goethite (GDS240), and hematite (GDS576) .
}
\references{
Clark, R. N., et al. (2007). "USGS digital spectral library splib06a." US geological survey, digital data series 231: 2007.

}
\examples{
data('mineralRef')
}
