\name{rutherglenNIR}
\alias{rutherglenNIR}
\title{Near-infrared subset of soil spectra from the Rutherglen area}
\usage{
data('rutherglenNIR')
}
\description{
Small set of soil spectra to be used for compartive purposes. The dataset is a dataframe with 12 spectra.}
\examples{
data('rutherglenNIR')
}
