\name{specSoilCol}
\alias{specSoilCol}
\title{Soil spectra with associated value of the horizon and soil type}
\usage{
data('specSoilCol')
}
\description{
A data frame with 2153 columns and 12 rows. The first two columns contains the soil type and horizon, respectively.
}
\examples{
## load the data into the environment
data('specSoilCol')
}
\details{The soil spectra and color dataset is a small set of vis-NIR spectra collected from a soil profile situated in the Rutherglen area during the Summer 2013. The vis-NIR spectra collected from the soil profile was done in-situ and measurements were made every 10 cm down to profile.}
