\name{datsoilspc}
\alias{datsoilspc}
\title{VIS-Nir data from Geeves Dataset}
\usage{
data('datsoilspc')
}
\description{
Extract of Geeves dataset including 391 observations in the NIR region with 5 soil attributes

\itemize{
   \item{clay: Clay content in percent}
   \item{silt: Silt content in percent}
   \item{sand: Sand content in percent}
   \item{TotalCarbon: Total carbon content in percent}
}}
\references{
Geeves, G. W. "physical, chemical and morphological properties of soils in the wheat-belt of southern NSW and northern Victoria." (1994).
}
\examples{
data('datsoilspc')
}
