\name{datsoilspc}
\alias{datsoilspc}
\title{VIS-Nir data from Geeves Dataset}
\usage{
data('datsoilspc')
}
\description{

Extract from data published by Geeves et al. (1994), The original dataset consists of 72 soil profiles, and a total of 424 soil samples from different depths in the wheat-belt of southern NSW and northern Victoria, Australia. The clay mineral fractions are mostly dominated by illite and kaolinite minerals. This dataset contains 391 observations in the NIR region with 4 measured soil attributes.

\itemize{
   \item{clay: Clay content in percent}
   \item{silt: Silt content in percent}
   \item{sand: Sand content in percent}
   \item{TotalCarbon: Total carbon content in percent}
}}
\references{
Geeves, G. W. "physical, chemical and morphological properties of soils in the wheat-belt of southern NSW and northern Victoria." (1994).
}
\examples{
data('datsoilspc')
}
