\name{datsoilspc}
\alias{datsoilspc}
\title{Vis-NIR infrared data}
\usage{
data('datsoilspc')
}
\description{
A data frame with 391 rows and 5 variables
\itemize{
   \item{clay: Clay content in percent}
   \item{silt: Silt content in percent}
   \item{sand: Sand content in percent}
   \item{Total_Carbon: Total carbon content in percent}
}}
\examples{
data('datsoilspc')
}
