% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra2colour.R
\name{spectra2colour}
\alias{spectra2colour}
\title{Method that extracts RGB and Munsell colour from a matri or dataframe of spectra.}
\usage{
spectra2colour(spectra, ...)
}
\arguments{
\item{spectra}{a matrix or dataframe containing the spectra}

\item{...}{other arguments to be passed in the function}
}
\description{
Converts spectra reflectance into RGB and Munsell colours
}
\values{
A dataframe with of size the number of bands plus two. The columns correspond to the number of spectral bands. The RBG colors are accessed by the column name `colour` and the Munsel color are accessed by the column name `munsell`. One color is provided for each spectra. The spectra is also returned.
}
