% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra2colour.R
\name{spectra2colour}
\alias{spectra2colour}
\title{Method that extracts RGB and Munsell colour from a matri or dataframe of spectra.}
\usage{
spectra2colour(spectra, ...)
}
\arguments{
\item{spectra}{a matrix or dataframe containing the spectra}

\item{...}{other arguments to be passed in the function}
}
\description{
Converts spectra reflectance into RGB and Munsell colours
}
\author{
Michael Nelson and Mario Fajardo
}
